Implement cache and request utilities
